requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django
django-debug-toolbar==4.2.0
django-environ==0.11.2
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.8.0
requests
djangorestframework
django-cors-headers
pyyaml
